Добавляет модули util и data
